Code refactoring: Reusing chart display
